waitingblock/requirements-dev.txt
Babel==2.6.0
Django==2.1.3
django-crispy-forms==1.7.2
django-phonenumber-field==2.1.0
django-tables2==2.0.2
phonenumbers==8.9.16
pytz==2018.7
yapf==0.24.0
djangorestframework==3.9.0
django-rest-framework-social-oauth2
Pillow==5.3.0
coreapi==2.3.3
PyYAML==3.13
